Refactor. Ball&Stick Highlighting
